requirements.txt
torch==2.7.0
transformers
unsloth
openai
requests
Pillow
opencv-python
python-dotenv
numpy
wandb
openai
python-dotenv
fal-client
boto3
opencv-python
numpy
anthropic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
ray[serve]
protobuf==3.20.3